+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 I will record changes to this file just so i don't need to look at commit history every time
 
 # Changelog
-2025-09-27: Just created this file. Everything before this date is in commit history.
+2025-09-27:
+- Just created this file. Everything before this date is in commit history.
+- Added kinto.py config file to repo.
+- Added homebrew to debian setup script.
+- Tweaked how vscode is installed on debian.